修改tinyint 为Integer类型，以前byte类型无法转换; 添加sink.table.sinkDate参数，设置默认入湖（库）时间字段

+++ dlink-metadata/dlink-metadata-mysql/src/main/java/com/dlink/metadata/convert/MySqlTypeConvert.java
@@ -18,7 +18,7 @@ public class MySqlTypeConvert implements ITypeConvert {
         }
         String t = column.getType().toLowerCase();
         if (t.contains("tinyint")) {
-            return ColumnType.BYTE;
+            return ColumnType.INTEGER;
         } else if (t.contains("smallint") || t.contains("tinyint unsigned")) {
             return ColumnType.SHORT;
         } else if (t.contains("bigint unsigned") || t.contains("numeric") || t.contains("decimal")) {
